First version of the register system, with a basic front-end

+++ src/main/java/br/com/estoqueBr/repository/EntradaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.estoqueBr.model.Entrada;
 
 @Repository
-public interface EntradaRepository extends JpaRepository<Entrada, Integer>{
+public interface EntradaRepository extends JpaRepository<Entrada, Long>{
 
 }
