+++ src/main/java/br/com/estoqueBr/repository/NotaFiscalRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.estoqueBr.model.NotaFiscal;
 
 @Repository
-public interface NotaFiscalRepository extends JpaRepository<NotaFiscal, Integer>{
+public interface NotaFiscalRepository extends JpaRepository<NotaFiscal, Long>{
 
 }
