+++ src/main/java/br/com/estoqueBr/repository/UnidadeMedidaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.estoqueBr.model.UnidadeMedida;
 
 @Repository
-public interface UnidadeMedidaRepository extends JpaRepository<UnidadeMedida, Integer> {
+public interface UnidadeMedidaRepository extends JpaRepository<UnidadeMedida, Long> {
 
 }
